requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.10.4
bcrypt==4.0.1
click==8.3.1
cryptography==44.0.2
fastapi==0.135.1
h11==0.16.0
httpx==0.28.1
idna==3.11
lxml==5.3.2
Mako==1.3.10
MarkupSafe==3.0.3
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
reportlab==4.4.10
resend==2.29.0
pdfplumber==0.11.8
